refactor links to use a seperate form that loads better
